8289230: Move PlatformXXX class declarations out of os_xxx.hpp
