+++ src/hotspot/os/posix/os_posix.hpp
@@ -96,195 +96,4 @@ public:
   static void print_active_locale(outputStream* st);
 };
 
-/*
- * Crash protection for the JfrSampler thread. Wrap the callback
- * with a sigsetjmp and in case of a SIGSEGV/SIGBUS we siglongjmp
- * back.
- * To be able to use this - don't take locks, don't rely on destructors,
- * don't make OS library calls, don't allocate memory, don't print,
- * don't call code that could leave the heap / memory in an inconsistent state,
- * or anything else where we are not in control if we suddenly jump out.
- */
-class ThreadCrashProtection : public StackObj {
-public:
-  static bool is_crash_protected(Thread* thr) {
-    return _crash_protection != NULL && _protected_thread == thr;
-  }
-
-  ThreadCrashProtection();
-  bool call(os::CrashProtectionCallback& cb);
-
-  static void check_crash_protection(int signal, Thread* thread);
-private:
-  static Thread* _protected_thread;
-  static ThreadCrashProtection* _crash_protection;
-  void restore();
-  sigjmp_buf _jmpbuf;
-};
-
-/*
- * This is the platform-specific implementation underpinning
- * the ParkEvent class, which itself underpins Java-level monitor
- * operations. See park.hpp for details.
- * These event objects are type-stable and immortal - we never delete them.
- * Events are associated with a thread for the lifetime of the thread.
- */
-class PlatformEvent : public CHeapObj<mtSynchronizer> {
- private:
-  double cachePad[4];        // Increase odds that _mutex is sole occupant of cache line
-  volatile int _event;       // Event count/permit: -1, 0 or 1
-  volatile int _nParked;     // Indicates if associated thread is blocked: 0 or 1
-  pthread_mutex_t _mutex[1]; // Native mutex for locking
-  pthread_cond_t  _cond[1];  // Native condition variable for blocking
-  double postPad[2];
-
- protected:       // TODO-FIXME: make dtor private
-  ~PlatformEvent() { guarantee(false, "invariant"); } // immortal so can't delete
-
- public:
-  PlatformEvent();
-  void park();
-  int  park(jlong millis);
-  void unpark();
-
-  // Use caution with reset() and fired() -- they may require MEMBARs
-  void reset() { _event = 0; }
-  int  fired() { return _event; }
-};
-
-// JSR166 support
-// PlatformParker provides the platform dependent base class for the
-// Parker class. It basically provides the internal data structures:
-// - mutex and convars
-// which are then used directly by the Parker methods defined in the OS
-// specific implementation files.
-// There is significant overlap between the funcionality supported in the
-// combination of Parker+PlatformParker and PlatformEvent (above). If Parker
-// were more like ObjectMonitor we could use PlatformEvent in both (with some
-// API updates of course). But Parker methods use fastpaths that break that
-// level of encapsulation - so combining the two remains a future project.
-
-class PlatformParker {
-  NONCOPYABLE(PlatformParker);
- protected:
-  enum {
-    REL_INDEX = 0,
-    ABS_INDEX = 1
-  };
-  volatile int _counter;
-  int _cur_index;  // which cond is in use: -1, 0, 1
-  pthread_mutex_t _mutex[1];
-  pthread_cond_t  _cond[2]; // one for relative times and one for absolute
-
- public:
-  PlatformParker();
-  ~PlatformParker();
-};
-
-// Workaround for a bug in macOSX kernel's pthread support (fixed in Mojave?).
-// Avoid ever allocating a pthread_mutex_t at the same address as one of our
-// former pthread_cond_t, by using freelists of mutexes and condvars.
-// Conditional to avoid extra indirection and padding loss on other platforms.
-#ifdef __APPLE__
-#define PLATFORM_MONITOR_IMPL_INDIRECT 1
-#else
-#define PLATFORM_MONITOR_IMPL_INDIRECT 0
-#endif
-
-// Platform specific implementations that underpin VM Mutex/Monitor classes.
-// Note that we use "normal" pthread_mutex_t attributes so that recursive
-// locking is not supported, which matches the expected semantics of the
-// VM Mutex class.
-
-class PlatformMutex : public CHeapObj<mtSynchronizer> {
-#if PLATFORM_MONITOR_IMPL_INDIRECT
-  class Mutex : public CHeapObj<mtSynchronizer> {
-   public:
-    pthread_mutex_t _mutex;
-    Mutex* _next;
-
-    Mutex();
-    ~Mutex();
-  };
-
-  Mutex* _impl;
-
-  static pthread_mutex_t _freelist_lock; // used for mutex and cond freelists
-  static Mutex* _mutex_freelist;
-
- protected:
-  class WithFreeListLocked;
-  pthread_mutex_t* mutex() { return &(_impl->_mutex); }
-
- public:
-  PlatformMutex();              // Use freelist allocation of impl.
-  ~PlatformMutex();
-
-  static void init();           // Initialize the freelist.
-
-#else
-
-  pthread_mutex_t _mutex;
-
- protected:
-  pthread_mutex_t* mutex() { return &_mutex; }
-
- public:
-  static void init() {}         // Nothing needed for the non-indirect case.
-
-  PlatformMutex();
-  ~PlatformMutex();
-
-#endif // PLATFORM_MONITOR_IMPL_INDIRECT
-
- private:
-  NONCOPYABLE(PlatformMutex);
-
- public:
-  void lock();
-  void unlock();
-  bool try_lock();
-};
-
-class PlatformMonitor : public PlatformMutex {
-#if PLATFORM_MONITOR_IMPL_INDIRECT
-  class Cond : public CHeapObj<mtSynchronizer> {
-   public:
-    pthread_cond_t _cond;
-    Cond* _next;
-
-    Cond();
-    ~Cond();
-  };
-
-  Cond* _impl;
-
-  static Cond* _cond_freelist;
-
-  pthread_cond_t* cond() { return &(_impl->_cond); }
-
- public:
-  PlatformMonitor();            // Use freelist allocation of impl.
-  ~PlatformMonitor();
-
-#else
-
-  pthread_cond_t _cond;
-  pthread_cond_t* cond() { return &_cond; }
-
- public:
-  PlatformMonitor();
-  ~PlatformMonitor();
-
-#endif // PLATFORM_MONITOR_IMPL_INDIRECT
-
- private:
-  NONCOPYABLE(PlatformMonitor);
-
- public:
-  int wait(jlong millis);
-  void notify();
-  void notify_all();
-};
-
 #endif // OS_POSIX_OS_POSIX_HPP
