+++ src/hotspot/os/windows/os_windows.hpp
@@ -133,99 +133,4 @@ public:
   static inline int get_thread_ptr_offset() { return _thread_ptr_offset; }
 };
 
-/*
- * Crash protection for the JfrSampler thread. Wrap the callback
- * with a __try { call() }
- * To be able to use this - don't take locks, don't rely on destructors,
- * don't make OS library calls, don't allocate memory, don't print,
- * don't call code that could leave the heap / memory in an inconsistent state,
- * or anything else where we are not in control if we suddenly jump out.
- */
-class ThreadCrashProtection : public StackObj {
-public:
-  static bool is_crash_protected(Thread* thr) {
-    return _crash_protection != NULL && _protected_thread == thr;
-  }
-
-  ThreadCrashProtection();
-  bool call(os::CrashProtectionCallback& cb);
-private:
-  static Thread* _protected_thread;
-  static ThreadCrashProtection* _crash_protection;
-};
-
-class PlatformEvent : public CHeapObj<mtSynchronizer> {
-  private:
-    double CachePad [4] ;   // increase odds that _Event is sole occupant of cache line
-    volatile int _Event ;
-    HANDLE _ParkHandle ;
-
-  public:       // TODO-FIXME: make dtor private
-    ~PlatformEvent() { guarantee (0, "invariant") ; }
-
-  public:
-    PlatformEvent() {
-      _Event   = 0 ;
-      _ParkHandle = CreateEvent (NULL, false, false, NULL) ;
-      guarantee (_ParkHandle != NULL, "invariant") ;
-    }
-
-    // Exercise caution using reset() and fired() - they may require MEMBARs
-    void reset() { _Event = 0 ; }
-    int  fired() { return _Event; }
-    void park () ;
-    void unpark () ;
-    int  park (jlong millis) ;
-} ;
-
-
-
-class PlatformParker {
-  NONCOPYABLE(PlatformParker);
-
- protected:
-  HANDLE _ParkHandle;
-
- public:
-  PlatformParker() {
-    _ParkHandle = CreateEvent (NULL, true, false, NULL) ;
-    guarantee(_ParkHandle != NULL, "invariant") ;
-  }
-  ~PlatformParker() {
-    CloseHandle(_ParkHandle);
-  }
-};
-
-// Platform specific implementations that underpin VM Mutex/Monitor classes.
-// Note that CRITICAL_SECTION supports recursive locking, while the semantics
-// of the VM Mutex class does not. It is up to the Mutex class to hide this
-// difference in behaviour.
-
-class PlatformMutex : public CHeapObj<mtSynchronizer> {
-  NONCOPYABLE(PlatformMutex);
-
- protected:
-  CRITICAL_SECTION   _mutex; // Native mutex for locking
-
- public:
-  PlatformMutex();
-  ~PlatformMutex();
-  void lock();
-  void unlock();
-  bool try_lock();
-};
-
-class PlatformMonitor : public PlatformMutex {
- private:
-  CONDITION_VARIABLE _cond;  // Native condition variable for blocking
-  NONCOPYABLE(PlatformMonitor);
-
- public:
-  PlatformMonitor();
-  ~PlatformMonitor();
-  int wait(jlong millis);
-  void notify();
-  void notify_all();
-};
-
 #endif // OS_WINDOWS_OS_WINDOWS_HPP
