+++ src/hotspot/os/windows/threadCrashProtection_windows.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,37 +22,33 @@
  *
  */
 
-#ifndef SHARE_RUNTIME_SEMAPHORE_HPP
-#define SHARE_RUNTIME_SEMAPHORE_HPP
+#ifndef OS_WINDOWS_THREADCRASHPROTECTION_WINDOWS_HPP
+#define OS_WINDOWS_THREADCRASHPROTECTION_WINDOWS_HPP
 
 #include "memory/allocation.hpp"
-#include "utilities/globalDefinitions.hpp"
 
-#if defined(LINUX) || defined(AIX)
-# include "semaphore_posix.hpp"
-#else
-# include OS_HEADER(semaphore)
-#endif
+class CrashProtectionCallback;
+class Thread;
 
-class JavaThread;
+/*
+ * Crash protection for the JfrSampler thread. Wrap the callback
+ * with a __try { call() }
+ * To be able to use this - don't take locks, don't rely on destructors,
+ * don't make OS library calls, don't allocate memory, don't print,
+ * don't call code that could leave the heap / memory in an inconsistent state,
+ * or anything else where we are not in control if we suddenly jump out.
+ */
+class ThreadCrashProtection : public StackObj {
+public:
+  static bool is_crash_protected(Thread* thr) {
+    return _crash_protection != NULL && _protected_thread == thr;
+  }
 
-// Implements the limited, platform independent Semaphore API.
-class Semaphore : public CHeapObj<mtSynchronizer> {
-  SemaphoreImpl _impl;
-
-  NONCOPYABLE(Semaphore);
-
- public:
-  Semaphore(uint value = 0) : _impl(value) {}
-  ~Semaphore() {}
-
-  void signal(uint count = 1) { _impl.signal(count); }
-
-  void wait()                 { _impl.wait(); }
-
-  bool trywait()              { return _impl.trywait(); }
-
-  void wait_with_safepoint_check(JavaThread* thread);
+  ThreadCrashProtection();
+  bool call(CrashProtectionCallback& cb);
+private:
+  static Thread* _protected_thread;
+  static ThreadCrashProtection* _crash_protection;
 };
 
-#endif // SHARE_RUNTIME_SEMAPHORE_HPP
+#endif // OS_WINDOWS_THREADCRASHPROTECTION_WINDOWS_HPP
