+++ src/hotspot/share/memory/metaspace/rootChunkArea.cpp
@@ -35,6 +35,7 @@
 #include "runtime/mutexLocker.hpp"
 #include "utilities/debug.hpp"
 #include "utilities/globalDefinitions.hpp"
+#include "utilities/ostream.hpp"
 
 namespace metaspace {
