+++ src/hotspot/share/runtime/threadCrashProtection.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2017, 2022, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -22,37 +22,21 @@
  *
  */
 
-#ifndef SHARE_RUNTIME_SEMAPHORE_HPP
-#define SHARE_RUNTIME_SEMAPHORE_HPP
+
+#ifndef SHARE_RUNTIME_THREADCRASHPROTECTION_HPP
+#define SHARE_RUNTIME_THREADCRASHPROTECTION_HPP
 
 #include "memory/allocation.hpp"
-#include "utilities/globalDefinitions.hpp"
 
-#if defined(LINUX) || defined(AIX)
-# include "semaphore_posix.hpp"
-#else
-# include OS_HEADER(semaphore)
-#endif
-
-class JavaThread;
-
-// Implements the limited, platform independent Semaphore API.
-class Semaphore : public CHeapObj<mtSynchronizer> {
-  SemaphoreImpl _impl;
-
-  NONCOPYABLE(Semaphore);
-
- public:
-  Semaphore(uint value = 0) : _impl(value) {}
-  ~Semaphore() {}
-
-  void signal(uint count = 1) { _impl.signal(count); }
-
-  void wait()                 { _impl.wait(); }
-
-  bool trywait()              { return _impl.trywait(); }
-
-  void wait_with_safepoint_check(JavaThread* thread);
+class CrashProtectionCallback : public StackObj {
+public:
+  virtual void call() = 0;
 };
 
-#endif // SHARE_RUNTIME_SEMAPHORE_HPP
+#if defined(LINUX) || defined(AIX) || defined(BSD)
+# include "threadCrashProtection_posix.hpp"
+#else
+# include OS_HEADER(threadCrashProtection)
+#endif
+
+#endif // SHARE_RUNTIME_THREADCRASHPROTECTION_HPP
